osd: fix printf warning on pg_log_entry_t::get_key_name
warning: osd/osd_types.cc:1716:76: format '%lu' expects argument of type 'long unsigned int', but argument 5 has type 'version_t {aka long long unsigned int}' [-Wformat]
warning: osd/osd_types.cc:1716:76: format '%lu' expects argument of type 'long unsigned int', but argument 5 has type 'version_t {aka long long unsigned int}' [-Wformat]

+++ src/osd/osd_types.cc
@@ -1713,7 +1713,7 @@ void pg_query_t::generate_test_instances(list<pg_query_t*>& o)
 string pg_log_entry_t::get_key_name() const
 {
   char key[40];
-  snprintf(key, sizeof(key), "%010u.%020lu", version.epoch, version.version);
+  snprintf(key, sizeof(key), "%010u.%020llu", version.epoch, (long long unsigned)version.version);
   return string(key);
 }
